+++ src/e2e/app/cli/inmem/tests/tests/commands/test_delete_command.rs
@@ -13,41 +13,21 @@ use kamu_cli_e2e_common::prelude::*;
 
 kamu_cli_execute_command_e2e_test!(
     storage = inmem,
-    fixture = kamu_cli_e2e_repo_tests::test_init_multi_tenant_creates_sqlite_database,
-    options = Options::default().with_no_workspace()
+    fixture = kamu_cli_e2e_repo_tests::test_delete_dataset
 );
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 kamu_cli_execute_command_e2e_test!(
     storage = inmem,
-    fixture =
-        kamu_cli_e2e_repo_tests::test_init_multi_tenant_with_exists_ok_flag_creates_sqlite_database,
-    options = Options::default().with_no_workspace()
+    fixture = kamu_cli_e2e_repo_tests::test_delete_dataset_recursive
 );
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 kamu_cli_execute_command_e2e_test!(
     storage = inmem,
-    fixture = kamu_cli_e2e_repo_tests::test_init_exist_ok_st,
-    options = Options::default().with_no_workspace()
-);
-
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-kamu_cli_execute_command_e2e_test!(
-    storage = inmem,
-    fixture = kamu_cli_e2e_repo_tests::test_init_exist_ok_mt,
-    options = Options::default().with_no_workspace()
-);
-
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-kamu_cli_execute_command_e2e_test!(
-    storage = inmem,
-    fixture = kamu_cli_e2e_repo_tests::test_init_in_an_existing_workspace,
-    options = Options::default().with_no_workspace()
+    fixture = kamu_cli_e2e_repo_tests::test_delete_dataset_all
 );
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////